backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
pytesseract 
pdfminer.six 
pillow